isGrey added, debugged; work on winning and losing, more bug?
bug found:
double click = flag (double click again cancels), which is gray, and cannot lose if click on it, but when single click on the flag, it reveals underneath, and gray stays. If cancels the flag, you may see the number underneath. If this number is -1 (bomb), you do not lose. (isGrey debugs this)
isGrey makes sure that if it is flagged, you may not reveal the number underneath (at least I hope)